+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 5.1.0 (IN PROGRESS)
 
+* Add donor info to Fund. Refs UIF-473.
+
 ## [5.0.0](https://github.com/folio-org/ui-finance/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v4.0.4...v5.0.0)